Create a component to produce a navbar
