+++ server/public/styles.css
@@ -31,6 +31,9 @@
 .center-content {
   display: flex;
   justify-content: center;
+}
+
+.full-width {
   width: 100%;
 }
 
@@ -59,14 +62,48 @@
   width: 105px;
 }
 
-.purple {
+.purple-fill {
   background: #9320e0;
 }
 
-.blue {
+.blue-fill {
   background: #3198fc;
+}
+
+.blue {
+  color: #3198fc;
+}
+
+.purple {
+  color: #9320e0;
 }
 
 .click {
   cursor: pointer;
+}
+
+.nav-bar {
+  background: rgb(12, 0, 83);
+  height: 45px;
+  padding: 5px;
+  display: flex;
+  justify-content: space-between;
+}
+
+.small-logo {
+  max-width: 200px;
+  height: 35px;
+}
+
+.user-circle {
+  height: 35px;
+  width: 35px;
+  border: 3px solid #ed105b;
+  border-radius: 100%;
+  align-items: center;
+  margin-left: 20px;
+}
+
+.weight-800 {
+  font-weight: 800;
 }
